Add support for query parameters.
We support querying by any of the fields in the dataset, so all dataset
columns are valid.  To this we add any parameters that are needed to
supplement the query semantics (e.g. limit).

Note that for now this currently requires users to quote query
parameters as appropriate for a pandas query; for example, when query
"weather":

```
  curl -w "\n%{http_code}\n" "localhost:8000/query?weather=%22rain%22"
```