Add test app as Docker image.
A test application is bundled as a Docker image.  `make` will build both
the `service` and `testclient` images.  With the service running using:

```
  docker run -t -p 80:8000 10x/weather-service:0.0.1
```

the test client will execute against the service using:

```
  docker run -t 10x/weather-testclient:0.0.1
```

Note that because the client is simply a test application, and not an
integral part of the overall application, it assumes that it is
conecting to a "local" service, and not using a multi-container
configuration as one might build using [compose].

[compose]: https://docs.docker.com/compose/